На все фрагменты добавлен TextView с описанием что за экран
