+++ app/src/main/java/ru/practicum/android/diploma/ui/fragments/FavouritesFragment.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.fragment
+package ru.practicum.android.diploma.ui.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import ru.practicum.android.diploma.R
 
-class FilterFragment: Fragment() {
+class FavouritesFragment: Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        return inflater.inflate(R.layout.fragment_filter, container, false)
+        return inflater.inflate(R.layout.fragment_favourites, container, false)
     }
 }
