+++ app/src/main/res/layout/activity_root.xml
@@ -16,14 +16,6 @@
         app:navGraph="@navigation/main_navigation_graph"
         tools:ignore="MissingConstraints" />
 
-    <TextView
-        android:id="@+id/buildConfigReadExampleTextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Hello World!"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+    <!--bottom_navigation_view-->
 
 </androidx.constraintlayout.widget.ConstraintLayout>
